Finished most of adminservice db integration
Updated the new event and patch event functions, routed the new. Updated server port to 5001, moved setup.js to the shared-db so that the admin and client can use it without duplicating code. Added everything but the return for the getevents function for the clientservice. Removed an extra server.js.